updating packer.json and install_java to use local jdk tarball if it exists

+++ scripts/install_java.sh
@@ -1,11 +1,17 @@
 #!/bin/bash
 pushd .
 
-apt update
-apt install libc6-i386
+apt-get update
+apt-get install libc6-i386
+
+if [[ ! -f '/tmp/jdk/jdk-8u121-linux-i586.tar.gz' ]]; then
+    cd /tmp/jdk
+    apt-get install curl
+    curl -L -O -H "Cookie: oraclelicense=accept-securebackup-cookie" -k "http://download.oracle.com/otn-pub/java/jdk/8u121-b13/e9e7ea248e2c4826b92b3f075a80e441/jdk-8u121-linux-i586.tar.gz"
+fi
 
 cd /usr/local/lib/
-curl -L -O -H "Cookie: oraclelicense=accept-securebackup-cookie" -k "http://download.oracle.com/otn-pub/java/jdk/8u121-b13/e9e7ea248e2c4826b92b3f075a80e441/jdk-8u121-linux-i586.tar.gz"
+tar xvzf '/tmp/jdk/jdk-8u121-linux-i586.tar.gz'
 
 ln -s $PWD/jdk1.8.0_121 /usr/local/lib/java
